requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
spacy
python-docx
pdfplumber
python-multipart